requirements.txt
# Robot & Camera
fairino
pyrealsense2

# VLM (Qwen2.5-VL-7B)
transformers>=4.40.0
accelerate
torch>=2.2.0
torchvision

# Object Detection
ultralytics>=8.1.0

# Vision & GUI
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# Utilities
scipy
scikit-learn
